+++ CHANGELOG.md
@@ -5,3 +5,20 @@
   - Consul command/profiling/syslog gathering
   - Vault commands/syslog gathering
   - Zip archive of all output
+
+## v0.0.2
+
+- Initial info command
+- Update documentation
+
+## v0.0.3
+
+- Initial upload command
+- Improved help output
+- Update documentation
+
+## v0.0.4
+
+- Additional info command output
+- More help output touches
+- Update documentation